refactor: Extract cleanup functions for better separation of concerns
- Move node reference cleanup utilities to @giselle-sdk/node-utils package
- Extract handleDeleteNodeCleanup and handleDeleteTriggerCleanup functions
- Follow existing pattern from copyNode implementation for better consistency
- Remove redundant conditional logic in deleteNode for cleaner code
