+++ packages/node-utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./default-name";
 export * from "./node-factories";
+export * from "./node-reference-cleanup";
